Solved all warning and improve UI
1. Solved all warning
2. Fix crash problem when running demo application using latest Xcode
7(7A220)
3. Fix minor bugs in low level, compatible new hardware and old
hardware (new hardware using iAP 2 protocol)

+++ libs/source/src/buffer.h
@@ -20,10 +20,10 @@ typedef struct ct_buf {
 	unsigned int		overrun;
 } ct_buf_t;
 
-extern void         ct_buf_init(ct_buf_t *, void *, size_t);
-extern void         ct_buf_set(ct_buf_t *, void *, size_t);
-extern int          ct_buf_get(ct_buf_t *, void *, size_t);
-extern int          ct_buf_put(ct_buf_t *, const void *, size_t);
+extern void         ct_buf_init(ct_buf_t *, void *, unsigned int);
+extern void         ct_buf_set(ct_buf_t *, void *, unsigned int);
+extern int          ct_buf_get(ct_buf_t *, void *, unsigned int);
+extern int          ct_buf_put(ct_buf_t *, const void *, unsigned int);
 extern int          ct_buf_putc(ct_buf_t *, int);
 extern unsigned int	ct_buf_avail(ct_buf_t *);
 extern void *		ct_buf_head(ct_buf_t *);
